+++ src/index.ts
@@ -3,3 +3,9 @@ export { createMytril } from './entry-bundler.js';
 
 // components
 export { default as App } from './components/app/app.svelte';
+
+// grids
+export { default as Container } from './components/grids/container/container.svelte';
+export { default as Row } from './components/grids/row/row.svelte';
+export { default as Col } from './components/grids/col/col.svelte';
+export { default as Spacer } from './components/grids/spacer/spacer.svelte';