+++ src/plugins/css/marker.ts
@@ -1,4 +1,4 @@
-import { cssParser, cssParserPalette } from './parser.js';
+import { cssParser, cssParserComponents, cssParserPalette } from './parser.js';
 
 const markerBreakpoint = {
 	start: '/* @breakpoints */',
@@ -10,7 +10,7 @@ export const cssMarkerPalette = (palette?: string) => {
 };
 
 export const cssMarker = (breakpoints: { [key: string]: number }): string => {
-	const css = cssParser();
+	const css = cssParser() + '\n' + cssParserComponents();
 	const startMarkerLength = markerBreakpoint.start.length;
 	const endMarkerLength = markerBreakpoint.end.length;
 
@@ -36,12 +36,12 @@ export const cssMarker = (breakpoints: { [key: string]: number }): string => {
 	let response = '';
 	if (extractedContent && extractedContent !== '') {
 		for (const property in breakpoints) {
-			if (response !== 'default') {
+			if (property !== 'default') {
 				response += `@media screen and (min-width: ${breakpoints[property]}px) {`;
 				response += extractedContent.trim().replaceAll('[breakpoint]', `.${property}\\:`);
 				response += `}`;
 				response += `@media screen and (max-width: ${breakpoints[property]}px) {`;
-				response += extractedContent.trim().replaceAll('[breakpoint]', `.${property}\\:`);
+				response += extractedContent.trim().replaceAll('[breakpoint]', `.max-${property}\\:`);
 				response += `}`;
 			} else {
 				response += extractedContent.trim().replaceAll('[breakpoint]', `.`);