+++ src/utils/dom.ts
@@ -1,0 +1,3 @@
+export function className(...args: Array<string | undefined>) {
+	return args.filter((cls: string | undefined) => !!cls).join(' ');
+}